InsnPro: This is a little program for auto create insn for gcc use.
InsnPro is writeen by java, eclipse. it's use an auto template to
gen an gcc's file. a.c is input file, and output file is the same
as a.c.

Template should be as a.c only with quote lines, and output will be
current a.c file.

With eclipse, builded jar file also works fine. If you use and jar file,
command will like this:
    java -jar InsnPro.jar inputfile > outputfile

Have fine. :)